Reverting RetentionStoreReader as an interface.
Adding AbstractRetentionStoreReader as a placeholder for code reuse.
