+++ krati-main/src/retention/java/krati/retention/RetentionStoreReader.java
@@ -35,39 +35,15 @@ import krati.store.StoreReader;
  * 08/23, 2011 - Created <br/>
  * 09/21, 2011 - Added interface StoreReader <br/>
  */
-public abstract class RetentionStoreReader<K, V> implements RetentionClient<K>, StoreReader<K, V> {
-    private final static Logger logger = Logger.getLogger(RetentionStoreReader.class);
-    
+public interface RetentionStoreReader<K, V> extends RetentionClient<K>, StoreReader<K, V> {
     /**
      * @return the data source of this RetentionStoreReader.
      */
     public abstract String getSource();
     
     /**
-     * Gets a number of value events starting from a give position in the Retention.
-     * The number of events is determined internally by the Retention and it is
-     * up to the batch size.   
-     * 
-     * @param pos - the retention position from where events will be read
-     * @param map - the result map (keys to value events) to fill in 
-     * @return the next position from where new events will be read.
+     * Convenience method. Composition of get(Position, List<Event>) and get.
+     * For a trivial implementation, look at extending AbstractRetentionStoreReader.
      */
-    public Position get(Position pos, Map<K, Event<V>> map) {
-        ArrayList<Event<K>> list = new ArrayList<Event<K>>(1000);
-        Position nextPos = get(pos, list);
-        
-        for(Event<K> evt : list) {
-            K key = evt.getValue();
-            if(key != null) {
-                try {
-                    V value = get(key);
-                    map.put(key, new SimpleEvent<V>(value, evt.getClock()));
-                } catch(Exception e) {
-                    logger.warn(e.getMessage());
-                }
-            }
-        }
-        
-        return nextPos;
-    }
+    public Position get(Position pos, Map<K, Event<V>> map);
 }
